Move get_http_response to asgi class to make it easier to override it
