+++ strawberry/asgi/__init__.py
@@ -1,14 +1,17 @@
 import asyncio
+import json
 import typing
 
+from starlette import status
 from starlette.requests import Request
-from starlette.responses import Response
+from starlette.responses import HTMLResponse, JSONResponse, PlainTextResponse, Response
 from starlette.types import Receive, Scope, Send
 from starlette.websockets import WebSocket, WebSocketDisconnect, WebSocketState
 
 from graphql import GraphQLError
 from graphql.error import format_error as format_graphql_error
 
+from strawberry.file_uploads.data import replace_placeholders_with_files
 from strawberry.http import GraphQLHTTPResponse, process_result
 from strawberry.types import ExecutionResult
 
@@ -24,7 +27,7 @@ from .constants import (
     GQL_START,
     GQL_STOP,
 )
-from .http import get_http_response
+from .utils import get_graphiql_html
 
 
 class GraphQL:
@@ -191,6 +194,72 @@ class GraphQL:
 
         return await websocket.send_json(data)
 
+    def get_graphiql_response(self) -> HTMLResponse:
+        html = get_graphiql_html()
+
+        return HTMLResponse(html)
+
+    async def get_http_response(
+        self,
+        request: Request,
+        execute: typing.Callable,
+        process_result: typing.Callable,
+        graphiql: bool,
+        root_value: typing.Optional[typing.Any],
+        context: typing.Optional[typing.Any],
+    ) -> Response:
+        if request.method == "GET":
+            if not graphiql:
+                return HTMLResponse(status_code=status.HTTP_404_NOT_FOUND)
+
+            return self.get_graphiql_response()
+
+        if request.method == "POST":
+            content_type = request.headers.get("Content-Type", "")
+            if "application/json" in content_type:
+                data = await request.json()
+            elif content_type.startswith("multipart/form-data"):
+                multipart_data = await request.form()
+                operations = json.loads(multipart_data.get("operations", "{}"))
+                files_map = json.loads(multipart_data.get("map", "{}"))
+
+                data = replace_placeholders_with_files(
+                    operations, files_map, multipart_data
+                )
+
+            else:
+                return PlainTextResponse(
+                    "Unsupported Media Type",
+                    status_code=status.HTTP_415_UNSUPPORTED_MEDIA_TYPE,
+                )
+        else:
+            return PlainTextResponse(
+                "Method Not Allowed",
+                status_code=status.HTTP_405_METHOD_NOT_ALLOWED,
+            )
+
+        try:
+            query = data["query"]
+            variables = data.get("variables")
+            operation_name = data.get("operationName")
+        except KeyError:
+            return PlainTextResponse(
+                "No GraphQL query found in the request",
+                status_code=status.HTTP_400_BAD_REQUEST,
+            )
+
+        result = await execute(
+            query,
+            variables=variables,
+            context=context,
+            operation_name=operation_name,
+            root_value=root_value,
+        )
+
+        response_data = await process_result(request=request, result=result)
+
+        return JSONResponse(response_data, status_code=status.HTTP_200_OK)
+
     async def handle_http(self, scope: Scope, receive: Receive, send: Send):
         request = Request(scope=scope, receive=receive)
         root_value = await self.get_root_value(request)
@@ -205,7 +274,7 @@ class GraphQL:
 
         context = await self.get_context(request=request, response=sub_response)
 
-        response = await get_http_response(
+        response = await self.get_http_response(
             request=request,
             execute=self.execute,
             process_result=self.process_result,
